+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Not released
 
 - Improve performance widget client side calculations [#88](https://github.com/CartoDB/carto-react-lib/pull/88)
+- Change BigQuery source type name from 'bq' to 'bigquery' [#97](https://github.com/CartoDB/carto-react-lib/pull/97)
 
 ## 1.0.0-beta14 (2021-02-08)